+++ internal/api/health.go
@@ -1,0 +1,30 @@
+package api
+
+import (
+	"net/http"
+
+	"github.com/muety/wakapi/helpers"
+)
+
+type HealthCheckResponse struct {
+	Version        string `json:"version"`
+	Name           string `json:"name"`
+	Description    string `json:"description"`
+	DatabaseStatus int    `json:"database_status"`
+}
+
+// HealthCheck endpoint indicates if the gotrue api service is available
+func (a *APIv1) HealthCheck(w http.ResponseWriter, r *http.Request) {
+	var dbStatus int
+	if sqlDb, err := a.db.DB(); err == nil {
+		if err := sqlDb.Ping(); err == nil {
+			dbStatus = 1
+		}
+	}
+	helpers.RespondJSON(w, r, http.StatusOK, HealthCheckResponse{
+		Version:        "0.0.1",
+		Name:           "Wakana",
+		Description:    "Wakana is an api for developer activity logs generated by IDE plugins",
+		DatabaseStatus: dbStatus,
+	})
+}
